pyproject.toml
[tool.poetry]
name = "medical-equipment-store"
version = "0.1.0"
description = ""
authors = ["Nicola Ibrahim <nicolaibrahim969@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "medical_equipment_store"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.6"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.5"
django-filter = "^22.1"
python-dotenv = "^0.21.1"
djangorestframework-jwt = "^1.11.0"
djangorestframework-simplejwt = "^5.2.2"
pillow = "^9.4.0"
cors = "^1.0.1"
pip = "^23.0"
install = "^1.3.5"
django-cors-headers = "^3.13.0"
django-split-settings = "^1.2.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pre-commit = "^3.0.4"
flake8 = "^6.0.0"
isort = "^5.12.0"
yapf = "^0.32.0"
colorlog = "^6.7.0"
drf-yasg = "^1.21.5"
django-debug-toolbar = "^4.0.0"
django-extensions = "^3.2.1"

[tool.black]
line-length = 119
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
multi_line_output = 5
line_length = 119

[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
